src/backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
flask
flask-sock
opencv-python
deepface